+++ libs/components/datetime/src/lib/modules/datepicker/datepicker-input-fuzzy.directive.ts
@@ -559,13 +559,10 @@ export class SkyFuzzyDatepickerInputDirective
     );
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
   // istanbul ignore next
   #onChange = (_: any) => {};
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
   // istanbul ignore next
   #onTouched = () => {};
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
   // istanbul ignore next
   #onValidatorChange = () => {};